requirements.txt
pyaes==1.6.1
pyasn1==0.6.1
rsa==4.9.1
Telethon==1.40.0
python-dotenv
logfire
